misc/ig/requirements.txt
dnspython==1.15.0
gevent==1.1.2
greenlet==0.4.10
lxml==3.6.4
requests==2.12.0
